update the current provider to take advantage of these last commits

+++ conversions/conversions.go
@@ -1,6 +1,6 @@
 // Package conversions will host all of our unit conversion functions
 package conversions
 
-func MsToKph(v float32) int {
-	return int((3600 * v) / 1000)
+func MsToKph(v float32) uint16 {
+	return uint16((3600 * v) / 1000)
 }
